refactor(dropdown): use colors from schemas
clean all base files from colors

Closes #828
