+++ src/components/combo/themes/light/header/combo-header.base.scss
@@ -1,5 +1,5 @@
 @use '../../../../../styles/utilities' as *;
-@use '../../../../dropdown/themes/light/dropdown-header.base' as *;
+@use '../../../../dropdown/themes/light/header/dropdown-header.base' as *;
 
 :host {
     --component-size: var(--ig-size, var(--ig-size-medium));
