--- src/components/dropdown/themes/light/dropdown-header.base.scss
@@ -1,21 +1,0 @@
-@use '../../../../styles/common/component';
-@use '../../../../styles/utilities' as *;
-
-$header-color: var(--header-color, color(secondary, 500)) !default;
-
-:host {
-    @include type-style('overline');
-    @include ellipsis();
-
-    display: block;
-    white-space: nowrap;
-    position: relative;
-    color: $header-color;
-    pointer-events: none;
-    padding-inline: pad-inline(4px, 8px, 16px);
-    padding-block: pad-block(2px, 4px, 8px);
-}
-
-::slotted(*) {
-    font: inherit;
-}
