--- src/components/dropdown/themes/light/dropdown-header.bootstrap.scss
@@ -1,5 +1,0 @@
-@use '../../../../styles/utilities' as *;
-
-:host {
-    color: color(gray, 700);
-}
