--- src/components/dropdown/themes/light/dropdown-header.fluent.scss
@@ -1,5 +1,0 @@
-@use '../../../../styles/utilities' as *;
-
-:host {
-    padding: pad(2px, 4px, 8px);
-}
