--- src/components/dropdown/themes/light/dropdown-item.bootstrap.scss
@@ -1,42 +1,0 @@
-@use 'sass:map';
-@use '../../../../styles/utilities' as *;
-@use '../../../../styles/themes/schemas/bootstrap/light' as *;
-
-$schema: map.get($light-bootstrap-schema, 'drop-down');
-$item-border-radius: border-radius(map.get($schema, 'item-border-radius')...);
-$theme: digest-schema(extend(
-    $schema,
-    (
-        item-border-radius: $item-border-radius
-    )
-));
-$item-color: var(--item-color, color(gray, 900)) !default;
-$hover-item-color: var(--hover-item-color, contrast-color(gray, 100)) !default;
-
-:host {
-    @include css-vars-from-theme($theme, 'ig-dropdown');
-
-    border-radius: var-get($theme, 'item-border-radius');
-    color: $item-color;
-}
-
-:host(:hover),
-:host([active]) {
-    background: color(gray, 100);
-    color: $hover-item-color;
-}
-
-:host([selected]) {
-    background: color(primary, 500);
-    color: contrast-color(primary, 600);
-}
-
-:host([selected]:hover),
-:host([selected][active]) {
-    background: color(primary, 600);
-    color: contrast-color(primary, 600);
-}
-
-:host([disabled]) {
-    color: color(gray, 400);
-}
