--- src/components/dropdown/themes/light/dropdown-item.fluent.scss
@@ -1,41 +1,0 @@
-@use 'sass:map';
-@use '../../../../styles/utilities' as *;
-@use '../../../../styles/themes/schemas/fluent/light' as *;
-
-$schema: map.get($light-fluent-schema, 'drop-down');
-$item-border-radius: border-radius(map.get($schema, 'item-border-radius')...);
-$theme: digest-schema(extend(
-    $schema,
-    (
-        item-border-radius: $item-border-radius
-    )
-));
-$selected-item-color: var(--selected-item-color, contrast-color(gray, 200)) !default;
-
-:host {
-    @include css-vars-from-theme($theme, 'ig-dropdown');
-
-    border-radius: var-get($theme, 'item-border-radius');
-    padding: pad(2px, 4px, 8px);
-}
-
-:host(:hover),
-:host([active]) {
-    background: color(gray, 100);
-    color: contrast-color(gray, 100);
-}
-
-:host([selected]) {
-    background: color(gray, 200);
-    color: $selected-item-color;
-}
-
-:host([selected]:hover),
-:host([selected][active]) {
-    background: color(gray, 200);
-    color: $selected-item-color;
-}
-
-:host([disabled]) {
-    color: color(gray, 400);
-}
