--- src/components/dropdown/themes/light/dropdown-item.indigo.scss
@@ -1,39 +1,0 @@
-@use 'sass:map';
-@use '../../../../styles/utilities' as *;
-@use '../../../../styles/themes/schemas/indigo/light' as *;
-
-$schema: map.get($light-indigo-schema, 'drop-down');
-$item-border-radius: border-radius(map.get($schema, 'item-border-radius')...);
-$theme: digest-schema(extend(
-    $schema,
-    (
-        item-border-radius: $item-border-radius
-    )
-));
-
-:host {
-    @include css-vars-from-theme($theme, 'ig-dropdown');
-
-    border-radius: var-get($theme, 'item-border-radius');
-    color: color(gray, 700);
-}
-
-:host(:hover),
-:host([active]) {
-    background: color(primary, 500, .15);
-}
-
-:host([selected]) {
-    background: color(primary, 500);
-    color: contrast-color(primary, 500);
-}
-
-:host([selected]:hover),
-:host([selected][active]) {
-    background: color(primary, 400);
-    color: contrast-color(primary, 400);
-}
-
-:host([disabled]) {
-    color: color(gray, 400);
-}
