+++ src/components/dropdown/themes/light/dropdown.base.scss
@@ -1,32 +1,15 @@
-@use 'sass:map';
 @use '../../../../styles/common/component';
 @use '../../../../styles/utilities' as *;
-@use '../../../../styles/themes/schemas/material/light' as *;
-
-$schema: map.get($light-material-schema, 'drop-down');
-$border-radius: border-radius(map.get($schema, 'border-radius')...);
-$theme: digest-schema(extend(
-    $schema,
-    (
-        border-radius: $border-radius
-    )
-));
 
 :host {
-    @include css-vars-from-theme($theme, 'ig-dropdown');
-
     display: flex;
     font-family: var(--ig-font-family);
 }
 
 [part='base'] {
-    border-radius: var-get($theme, 'border-radius');
     overflow-y: auto;
     overflow-x: hidden;
-    background: color(surface, 500);
-    box-shadow: var(--ig-elevation-8);
     min-width: rem(112px);
-    padding: rem(8px) 0;
     z-index: 10005;
 }
