+++ src/components/dropdown/themes/light/group/dropdown-group.base.scss
@@ -1,5 +1,5 @@
-@use '../../../../styles/common/component';
-@use '../../../../styles/utilities' as *;
+@use '../../../../../styles/common/component';
+@use '../../../../../styles/utilities/index' as *;
 
 :host {
     pointer-events: auto;
@@ -11,7 +11,6 @@
         display: block;
         white-space: nowrap;
         position: relative;
-        color: color(secondary, 500);
         pointer-events: none;
         padding-inline: pad-inline(4px, 8px, 16px);
         padding-block: pad-block(2px, 4px, 8px);
