+++ src/components/dropdown/themes/light/group/dropdown-group.fluent.scss
@@ -1,6 +1,6 @@
 @use 'sass:map';
-@use '../../../../styles/utilities' as *;
-@use '../../../../styles/themes/schemas/fluent/light' as *;
+@use '../../../../../styles/utilities' as *;
+@use '../../../../../styles/themes/schemas/fluent/light' as *;
 
 $schema: map.get($light-fluent-schema, 'drop-down');
 $border-radius: border-radius(map.get($schema, 'border-radius')...);
@@ -13,10 +13,9 @@ $theme: digest-schema(extend(
 
 :host {
     @include css-vars-from-theme($theme, 'ig-dropdown');
-}
 
-[part='base'] {
-    background: color(surface, 500);
-    border-radius: var-get($theme, 'border-radius');
-    box-shadow: var(--ig-elevation-4);
+    ::slotted([slot='label']) {
+        color: color(secondary, 500);
+        padding: pad(2px, 4px, 8px);
+    }
 }
