+++ src/components/dropdown/themes/light/header/dropdown-header.base.scss
@@ -1,0 +1,16 @@
+@use '../../../../../styles/common/component';
+@use '../../../../../styles/utilities/index' as *;
+
+:host {
+    @include type-style('overline');
+    @include ellipsis();
+
+    display: block;
+    white-space: nowrap;
+    position: relative;
+    pointer-events: none;
+}
+
+::slotted(*) {
+    font: inherit;
+}
