+++ src/components/dropdown/themes/light/header/dropdown-header.indigo.scss
@@ -1,6 +1,6 @@
 @use 'sass:map';
-@use '../../../../styles/utilities' as *;
-@use '../../../../styles/themes/schemas/indigo/light' as *;
+@use '../../../../../styles/utilities' as *;
+@use '../../../../../styles/themes/schemas/indigo/light' as *;
 
 $schema: map.get($light-indigo-schema, 'drop-down');
 $border-radius: border-radius(map.get($schema, 'border-radius')...);
@@ -13,10 +13,8 @@ $theme: digest-schema(extend(
 
 :host {
     @include css-vars-from-theme($theme, 'ig-dropdown');
-}
 
-[part='base'] {
-    background: color(surface, 500);
-    border-radius: var-get($theme, 'border-radius');
-    box-shadow: var(--ig-elevation-3);
+    color: var(--header-color, color(secondary, 500));
+    padding-inline: pad-inline(4px, 8px, 16px);
+    padding-block: pad-block(2px, 4px, 8px);
 }
