+++ src/components/dropdown/themes/light/header/dropdown-header.material.scss
@@ -1,0 +1,20 @@
+@use 'sass:map';
+@use '../../../../../styles/utilities' as *;
+@use '../../../../../styles/themes/schemas/material/light' as *;
+
+$schema: map.get($light-material-schema, 'drop-down');
+$border-radius: border-radius(map.get($schema, 'border-radius')...);
+$theme: digest-schema(extend(
+    $schema,
+    (
+        border-radius: $border-radius
+    )
+));
+
+:host {
+    @include css-vars-from-theme($theme, 'ig-dropdown');
+
+    color: var(--header-color, color(secondary, 500));
+    padding-inline: pad-inline(4px, 8px, 16px);
+    padding-block: pad-block(2px, 4px, 8px);
+}
